api/requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
coverage==6.0.1
Flask==2.0.2
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-flask==1.2.0
python-dotenv==0.19.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
